pyproject.toml
[tool.poetry]
name = "pythonproject1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
forex-python = "^1.8"
types-requests = "^2.31.0.10"
python-dotenv = "^1.0.0"
logging = "^0.4.9.6"
pandas = "^2.1.3"
chardet = "^5.2.0"
openpyxl = "^3.1.2"


[tool.poetry.group.lint.dependencies]
flake8 = "^6.1.0"
black = "^23.10.0"
isort = "^5.12.0"
mypy = "^1.6.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py
)
'''

[tool.isort]
line_length = 119

[tool.mypy]
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'
ignore_missing_imports = true